added google authentication, edited flexbox
