+++ src/Components/Header.js
@@ -1,0 +1,16 @@
+import SignIn from "./SignIn";
+
+function Header() {
+  return (
+    <>
+        <SignIn />
+        <div className="app">
+        <h1>Fernweh</h1>
+        <h2>/FEIRN-veyh/ - noun</h2>
+        <h3>A longing for distant places</h3>
+      </div>
+    </>
+  );
+}
+
+export default Header;
